Create matrix3
In this code we take input of number of rows and columns for two matrix. We check whether number of columns of first matrix is equal to number of rows of second matrix. If it is not equal we say it is invalid, and take the input again. If the input is valid then we take input of the matrix. Add the two matrix, and store in a third matrix. Then we print the third matrix. For taking input and giving output of matrix we use for loop.